prepare for lmdb-load
